Change something with the distance sensor because its a double where it can be an int, also add color sensing

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/main/utils/io/OutputSpace.java
@@ -43,7 +43,7 @@ public class OutputSpace {
         return ELEVATOR_BOTTOM_LIMIT_SWITCH.returnOutput(type);
     }
 
-    public double receiveOutputFromHandDistanceSensor() {
+    public int receiveOutputFromHandDistanceSensor() {
         return HAND_DISTANCE_SENSOR.returnOutput();
     }
